runtime: an enum class could not be registered as an enumeration
ibValueEnumerationVariant::GetNumber returned the member by implicit conversion,
which only an unscoped enum does. Registering a scoped one therefore failed to
compile inside the template, on a line two functions below the one that already
casts (IsEmpty). A member number is a number; said with a cast, it is one for
every enumeration.

+++ src/engine/backend/compiler/enumUnit.h
@@ -163,7 +163,12 @@ protected:
 		}
 		// The identifier, for whoever needs it as such.
 		const wxString& GetEnumMemberName() const { return m_name; }
-		virtual ibNumber GetNumber() const override { return m_value; }
+		// ⚠ CAST, NOT AN IMPLICIT CONVERSION — the same one IsEmpty above already writes. A member
+		// number is a number, but only an UNSCOPED enum says so on its own: a scoped one
+		// (`enum class`) converts to nothing implicitly, so this line was what refused to register a
+		// scoped enumeration in the runtime at all — and refused it in the template, where the
+		// enumeration itself is fine.
+		virtual ibNumber GetNumber() const override { return static_cast<long>(m_value); }
 
 	private:
 		wxString m_name;
